NewMetricFromURLPath implementation + tests
There is a concern about 'Hash' field being public and therefore
inadvertently modifiable by external actors (e.g. another internal package).
The problem is that we need to render this field to JSON and in the future
store it in the database as well, so it needs to be public for
'json.Marshal' to discover it (but that's only if the call to
'json.Marshal' happens in another package).
Alternatively, we could implement 'json.Marshaller' interface to serialize
private (non-exported) fields, or provide a method for a 'Metric' model
that would handle serialization (e.g. 'ToJSON()').
This would work for the JSON encoding.
For other serialisation protocols, another solution might be needed.