agent: basic interfaces + tests (failing as expected)
Here we define basic interfaces and structs with dummy methods that
will be implemented in the next commits.
This commit also defines tests which are failing at the moment;
this is expected (in the spirit of TDD).
Next commits will be focused on making the tests green again
(and potentially fixing/expanding the tests too).