agent: main loop implementation
Collecting and storing metrics is done synchronously as those operations
are expected to finish quickly - at least for now: the metrics we are
collecting are readily available to the agent process and do not involve
any I/O or network calls.

Reporting, however, is performed asynchronously, in a separate goroutine,
because there is network call involved and there might be timeouts, etc.
When agent context is cancelled (e.g by a SIGINT), main loop will wait
for all still running goroutines to finish and report their errors,
then it will stop the execution.

To ensure proper termination of reporting goroutines, we should pass
proper context to the reporter itself and ensure that it uses
this context for HTTP requests,
thus ensuring their cancellation should the context be cancelled.