agent: metrics source interface + tests
To add more flexibility to our metrics collector, it was decided to
abstract metrics source with an interface and support multiple sources
in the collector itself.

This commit introduces 'Source' interface and tests covering
dummy implementation.
The tests are failing as expected - this is to be fixed in the next
commits.