improve router tests
Now using proper HTTP server (from httptest) and sample HTTP requests
to validate that correct handlers are responding.
This overlaps a bit with tests for a concrete handler, but only
covers a happy path here as the goal is to demonstrate that
correct handler is picked by the router,
while in the tests for the handler we cover more edge cases.