model: introduce metric.Copy method
There is a subtle bug in the 'service.Metrics' - whenever we change a
delta for a counter to include previously stored value, the original
metric object also changes because it contains pointers to the actual
values.
To work around that, we need to copy original metric object before
changing its value.
Tests confirming the bug are also added.