agent: serialize reporter calls
If reporter is being slow for some reason, we do not want to spawn
too many goroutines trying to send probably the same values to the
server - only goroutine should do the reporting, and when it finishes,
the next can proceed.
While a proper way to addressing this would be running reporter loop in
a separate "worker" goroutine, it would require too many changes in the
current "architecture", so we chose a simpler approach which requires
minimal changes: add a mutex and only proceed with reporting if mutex
is grabbed, otherwise new goroutine would return immediately.
Necessary tests are included to verify the solution.