agent: storer should store only last value for counters
Given that we always add an incoming value for a counter to its previous
value on the server side, it does not make sense to do the same on the
agent side, because it would lead to inflated counter values on the
server.
For example, if we take 'PollCount' metric (which is the only counter
being reported for now), in the previous implementation the agent would
store values as 1, 3, 6, 10, ... at each poll interval, and the server
would store them as 1, 4, 10, 20 ...

To be fair, the current implementation of the storer gives little
advantage over using 'repository.Storage' directly and might seem an
unnecessary abstraction, still, we shall keep it for now in case
future requirements might justify it.
If that would not be the case, we will dismantle it at a later stage.

There is another potential problem with reporting counters to the server.
Even if we report the last value of the counter from the agent,
the server would still add this value to the previously stored one -
this might or might not be incorrect from the logical point of view.
One way to tackle that would be making reporter store the last value sent
to the server and only send deltas on subsequent reports
(only for counters, that is).
Still, on the first start of the agent, the first value reported would
again be incorrect if the server already had a previous value stored -
the agent would report the current value of the counter, and the server
would add that to the existing one.
Although, this would not be a problem for a 'PollCount' as it would
start from '0', but if it was a system counter that was never reset to
'0', and agent was restarted, the initial reported value would be
the last value of that counter, not the delta.
To mitigate that, we can either use a persistent storage for the reporter,
or retrieve the last known value from the server itself upon the first
report.