agent: report only delta for counters
Here we attempt to make reporting of counters more accurate by reporting
only delta values instead of full values.

Why this matters?
Previously, the agent would store the last value of the counter and
report is as is to the server.
The server would add received value to the previously stored value, thus
inflating the real value of the counter.
For example, for 'PollCount' metric (the only counter for now), the
agent would report the following sequence
(assuming poll interval == report interval): 1, 2, 3, 4, 5, ...
The server would store another sequence: 1, 3, 6, 10, 15, ...
This does not seem correct from the logical point of view: the actual
number of polls by the agent in the example above would be 5,
but the server will store 15, not 5, which does not make sense.

To mitigate that, we use an internal storage (currently in-memory)
in the reporter itself, where it stores successfully sent metrics as is.
Whenever next round of reporting comes, the reporter would calculate
a delta for a given counter metric, and send the delta instead of the
full value.
This approach is not without issues, though.
For example, with current in-memory storage, whenever the agent is
restarted, it would send the full value on the first report.
If storage is not in-memory and cannot store the last value after
being successfully sent to the server, the next time the reporter
will report incorrect delta (or even full value).
A better approach might be having the reporter ask the server for the
last value on the first report and use it as a base to calculate a delta.
That said, we will stick with the current imperfect implementation
for the time being until further requirements emerge.