agent: dispose of unnecessary Storer intermediary
When we introduced an internal storage to the reporter, the role of
'Storer' component in the agent pipeline had diminished and become
unnecessary (as it was simply a proxy to the underlying storage).
Therefore, it appeared to make more sense to get rid of this component
in favor of having internal storage in the collector itself,
where it could store intermediate results of the polling.
The implementation of that required changing 'Collector' interface
to introduce 'Snapshot()' method, which would return all metrics from
the collector's internal storage (similar to what 'Storer.Retrieve()' did).