service: refine Metrics interface to segregate single/batch operations
(based on Code Review comments)

Instead of using a single method to store either single or multiple
metrics depending on the number of arguments passed, it was decided
to define separate methods for each type of operation.
This way, when we want a single metric, we would call `StoreSingle`
or `RetrieveSingle` method and get either a metric or error in return.
For multiple metrics, we would use `StoreBatch`/`RetrieveBatch` methods
explicitly as we would need to handle returned values differently.
E.g. `RetrieveBatch` would only return existing metrics without
throwing an error by silently skipping non-existent keys.
Finally, it was decided to keep `RetrieveAll` as a separate method
instead of generalising `RetrieveBatch` to return all metrics
if no specific keys were requested - this way a caller would express
a specific intent.

PS. We could've kept original `Store`/`Retrieve` names for methods
operating on a single metric/key, but we opted to provide more explicit
naming instead.