agent: decompose 'Run' entrypoint
Extract a generic function that can run a periodic task at a given interval
and optionally with a custom initial delay.
Now, 'Run' function can launch 2 goroutines (collector and reporter)
that utilise that periodic function and waits for them to finish
before exiting.
Graceful shutdown has moved from the 'Run' function into the
periodic function which watches for context cancellations and terminates
when the last chunk of work has completed.