reporter: decompose 'reportSingle' method
There are 3 independent parts in that function: getting adjusted
(sendable) metric value (as we only send deltas to maintain correct
values on the server side), sending metric to the server, and storing
metric that was successfully sent, to the internal storage.
Out of these parts, the first two are easily extractable into separate
methods, while the third part is a one-liner, so it stays inside
'reportSingle' function.
To be fair, there is quire a bit of copying going on for the 'Metric'
struct, so at some point we might need to get a pprof and switch to using
pointers if the pprof data hints at that. Or change internal
representation of the metric to something more performant.