logging: create an implementation-agnostic logging API
Motivation for this is described in more detail in the comments
inside `log.go` file.

TL;DR
Since this is a learning project, it seemed natural to attempt a creation
of an abstract logging API.
As a side effect of this effort, deeper understanding was gained into
the internal of various logging implementations.