logging: a special logger for tests and "noop" logger
Our internal logging package is getting more complex due to the need
to use it in tests for other packages.
At such rate, we might end up re-implementing functionality already
provided by existing logging framework (:facepalm:).
The good news is, this has been a quite helpful exercise in terms
of learning; on a real project though, we would stick with some
battle-hardened option, like `zap` :)

That said, here we go: a logger to use in tests, which will record
logging events and allow their inspection. It sort of mimicks what
`zap` already does with its `observer` subpackage
(and we could've used that, but that wouldn't be fun).
And while we were at it, we also found some bugs in our `zerolog`
implementation (not that we plan to use it anyway).
Also, a bunch of mutexes to protect vulnerable slices and maps
from potential data races - not that we could reproduce that yet,
but better be safe than sorry :)