service: proper handling of empty metrics
Empty metrics should not be passed to the underlying storage,
nor retrieved from it (the storage should not store such metrics itself).
By empty metric here we understand a `Metric` struct without both `Delta`
and `Value` fields (provided `ID` and `Type` fields are set).

Also, we break free from dependency on `repository.MemStorage` in
service tests and instead implement a `mockStorage` that works similarly
to `MemStorage` but is controlled by the tests themselves.