server: introducing metric snapshotter
Metric snapshotter implements `MetricStorer` interface and proxies
writes to the wrapped instance of the metric storer while recording
the fact of the writes happening.
On top of that, the snapshotter adds two more methods:
- `DumpClose` dumps all previously seen metrics to a provided writer
  and closes the writer on success;
- `LoadClose` loads all metrics from a provided reader, feeds them to
the underlying storer instance and closes the reader on success;

Additionally, the snapshotter exposes a buffered channel with the
capacity of 1 to facilitate notification of external actors when writes
happen.

This commit covers a high-level API sketch, some basic/straightforward
implementation and test stubs.
In the next commits, we will expand testing coverage and flesh out
the implementation details.