server: avoid truncating metrics store file too early
Metric snapshotter will not write anything to the provided writer
when no metrics were updated since the previous call to `DumpClose`
method.
And because we are truncating output file when opening it, all previous
metrics become lost.
One of the way to remedy that without rewriting the metric snapshotter
would be writing to a temporary file and checking its size - if it is
not empty, we rename it to become the original output file.

Also, the server should continue with its operations even if it failed
to load metrics on startup.