refactoring: extract 'router' package
Fair warning: this is quite a big commit that just shuffles a bunch of
code around (yes, tests got rewritten too).

A main motivation for this became a need to add more handlers, and,
consequently, more tests for the router.
The previous code structure was created in such a way that router
tests duplicated a subset of handlers' tests.
While it worked alright, it turned out not to be sustainable when
new headers needed to be added.

By extracting router functionality into a separate package and using
mock handlers, we can achieve more stable tests and prevent leaky
abstractions.
There were two options considered to achieve the separation of concerns:
1. Create a handler factory that would be passed to the router and
   the router will map routes by calling corresponding methods on
   the factory.
2. Create a handler registry (essentially a map) and assign to each
   handler an identifier that the router will use when constructing
   routes.

The 2nd option was chosen as slightly simpler in the implementation
and in the mocking.
Whether the router should be the one that maps routes to handlers is
another good question, but for now we've decided to keep it that way.
Otherwise, we could've just thrown the router away and use `chi.Router`
directly while defining routes somewhere in the app or in the handlers
themselves.