refactoring: add logger to all handlers
It is quite handy to have a logger readily available inside any handler.
This could be solved by adding a `logger` field to each handler or
by embedding a `baseHandler` with the `logger` field.
The latter approach has been chosen as it potentially allows to embed
more shared things into each handler, should the need arise.