app: add tests for 'app/agent' and 'app/server' packages
To be fair, those tests largely replicate the tests we already have
in `cmd/agent` and `cmd/server` main packages, but here we are not
limited by command line flags parsing and can tweak app config
to a greater degree.
Eventually, these tests should evolve into integration tests for the
app, and tests from `cmd/...` should probably become either end-to-end
tests or obsolete and be removed.