Add initial DB migrations and make server apply them at startup
It has been decided to use `goose` as the DB migration tool due to
easy embedding into the app. 'Goose' also supports Go-based
migrations for complex cases, should we need that in the future.

At any rate, we create two separate tables: one for counters and one
for gauges; there are no other types yet.
We have settled on using `varchar` as primary keys because there is
very little incentive to use surrogate integer primary keys:
- there are no other tables referencing metric ids;
- when updating/retrieving values, we need to lookup metric id anyway,
  which is a string, so string comparison will happen anyway;

In any case, it would be possible to migrate the data to a new schema
at a later stage, when new requirements for the project arise.

Testing coverage is not yet complete, but handles some essential cases.