refactoring: add 'Context' to 'Storage' and 'MetricStorer'
Now that we use a networked storage (PostgreSQL), it is important to
have a way to cancel long-running requests.
To achieve that goal, we have added a `context.Context` parameter to all
methods of `repository.Storage` and `service.MetricStorer` interfaces.
While doing that required quite a bit of refactoring, it will provide
us with numerous benefits in the future, e.g. canceling long-running
requests, proper graceful shutdown propagated to the service/storage
level, call tracing.