repository: SQL storage implementation with unit tests
Initial implementation of the SQL storage.
The unit tests use `sqlmock` library to stay fast and avoid launching
inevitable slow external Postgres process.
This approach is not without downsides, as it heavily relies on SQL
queries matching a pattern, thus making tests brittle and sensitive
to internal changes.
Still, this is our first line of defense to verify error handling and
internal logic based on a predefined set of data.

That said, this approach should be complemented with integration tests
which launch real Postgres database and perform queries against it.
This will be addressed in the following commits as we integrate our
SQL storage into the app logic.