model: repurpose 'MetricSet' for aggregation
Previous, `MetricSet` represented a slice of `Metric` objects,
but the name was misleading as `Set` implies some uniqueness.
Given that there were no active usages of `MetricSet` in the codebase,
we repurposed it to represent a unique set of `Metric` objects.
Since there is not a set data structure in Go stdlib, we are using
a map keyed by `MetricKey`.

Also, on top of simple aggregation strategy where last seen value wins,
we added a strategy that accumulates delta values for counter metrics.
This might become useful in the `MetricStorer` for batch updates.