repository: SQL storage redesigned
SQL storage now has `GetMulti` and `SetMulti` methods which can be
used for batch updates.
`SetMulti` performs inserts/updates in a transaction per metric type
since counters and gauges are stored in different tables and there
does not seem to be a reason to lock both tables when updating only one
of them.