Introducing batch writes for metric storer service
Fairly big commit here, as things got a bit out of control :)
The main idea was to add support for batch updates to the `MetricStorer`
without exposing too much details about the storage itself.
The resulting design is not perfect as we now need to remember to
launch our storage batch writer goroutine, and exposing its launch up to
the server level is probably not a good idea.
On the other hand, launching it silently from the metric storer itself
is also not without downsides, but at least it would have not required
a separate launch of the storage batch writer in tests.
Which brings us to the next question about the quality of the tests
in this project, and, indeed, there seem to be an excessive exposure
to the internals of a package under test.
We should probably focus more on the tests of the package public API
and integration tests, rather than drowning in the multitude of brittle
tests just for the sake of coverage...