refactoring: move 'log' and 'periodictask' to '/pkg'
Those a helper packages that are shared across all layers, so they
should live in place designated just for that.
Typically, this is either `/pkg` or `/internal/pkg` as per various
discussions in the community.

For these two particular modules, it seems to make more sense to place
them in the `/pkg` rather than creating `/internal/pkg`.