agent: attempting to fix flaky tests
It is hard to test concurrent code reliably - tests based on timings
are very sensitive to a hardware load, and Go scheduler adds more
"spice" with `-race` and `-parallel`.
In Go 1.25, there's a new `testing/synctest` package which is supposed
to help with such scenarios, but we're on Go 1.24 at the moment.
Go 1.24 also has this package, but it is marked experimental and requires
a special environment variable to use it.
This is not very convenient, given that we would like to avoid changing
Github actions pipeline.
So far, the only method available here is to tweak timeouts and hope
for the best (aka keep retrying until tests succeed).