hmacsigner: helper package for generating HMAC-SHA256 signatures
A note on `LoadSecretKey` function:

Yandex.Practicum autotests pass a file path as a key, but that file
is not created by autotests, so we treat provided secret key
literally, as a sequence of bytes.
Also, since autotests neither sign their requests nor verify server
responses (at least as of 2025-11-07), we could have as well ignore
non-existent file and behave as if no key was passed at all.
What we cannot do, however, is to fail with an error, because we would
fail autotests if we did.