handler: add HMAC signer middleware
When a secret key is present, the server will verify signatures of the
incoming requests if they have a non-empty hash header.
The server will reject incoming requests with incorrect signature
(i.e. body does not match the signature).
The server should be rejecting requests without hash header when the
secret key is configured, but, unfortunately, this will break current
Yandex.Practicum autotests (which neither sign their requests nor
verify responses' signature).

If request is accepted, then a necessary handler is invoked and its
response body and status code are analyzed in the middleware.
If status code < 300 and body is not empty, the server signs its
response using provided secret key.

When the secret key is not configured, the server does not interfere
with the request/response flow, which is the default behavior.