agent: new config option to set sender workers pool size
Default value of `0` mean that workers are disabled and any sending
is done serially (matching previous behavior).
Whenever pool size is set to a positive number, this number of workers
are created to send collected metrics in parallel.
The way metrics are divided between the workers will be decided by
the reporter (most likely in batches in 10 - could be made configurable).